+++ resources/js/Components/Global/AppVersion/AppVersion.vue
@@ -11,10 +11,10 @@
       TV &#169; 2009-{{ new Date().getFullYear() }}
     </div>
     <div class="flex justify-center">
-      Beta Preview v0.9.3.11
+      Beta Preview v0.9.3.12
     </div>
     <div class="flex justify-center text-blue-600 hover:text-blue-400 underline">
-      <Link @click="appSettingStore.closeNavDropdown()" :href="route('changelog.show')">changelog</Link>
+      <Link @click="appSettingStore.closeNavDropdown()" :href="`/changelog`">changelog</Link>
     </div>
     <!--                                        <div class="flex justify-center">Please send us</div>-->
     <!--                                          <div class="flex justify-center">comments and questions</div>-->